[Windows] Move Win32 new functions of CWIN32Util to new CDisplayUtilsWin32
