+++ xbmc/platform/win32/WIN32Util.cpp
@@ -25,11 +25,8 @@
 
 #include "platform/win32/CharsetConverter.h"
 
-// @todo: Remove this and "SDK_26100.h" when Windows SDK updated to 10.0.26100.0 in builders
 #ifdef TARGET_WINDOWS_DESKTOP
-#ifndef NTDDI_WIN11_GE // Windows SDK 10.0.26100.0 or newer
-#include "SDK_26100.h"
-#endif
+#include "DisplayUtilsWin32.h"
 #endif
 
 #include <PowrProf.h>
@@ -1205,75 +1202,6 @@ bool CWIN32Util::SetThreadLocalLocale(bool enable /* = true */)
   return _configthreadlocale(param) != -1;
 }
 
-#ifdef TARGET_WINDOWS_DESKTOP
-std::wstring CWIN32Util::GetCurrentDisplayName()
-{
-  MONITORINFOEXW mi{};
-  mi.cbSize = sizeof(mi);
-  if (!GetMonitorInfoW(MonitorFromWindow(g_hWnd, MONITOR_DEFAULTTOPRIMARY), &mi))
-    return {};
-  return mi.szDevice;
-}
-
-std::vector<DISPLAYCONFIG_PATH_INFO> CWIN32Util::GetDisplayConfigPaths()
-{
-  uint32_t pathCount{0};
-  uint32_t modeCount{0};
-  std::vector<DISPLAYCONFIG_PATH_INFO> paths;
-  std::vector<DISPLAYCONFIG_MODE_INFO> modes;
-
-  constexpr uint32_t flags{QDC_ONLY_ACTIVE_PATHS};
-  LONG result{ERROR_SUCCESS};
-
-  do
-  {
-    if (GetDisplayConfigBufferSizes(flags, &pathCount, &modeCount) != ERROR_SUCCESS)
-      return {};
-
-    paths.resize(pathCount);
-    modes.resize(modeCount);
-
-    result = QueryDisplayConfig(flags, &pathCount, paths.data(), &modeCount, modes.data(), nullptr);
-  } while (result == ERROR_INSUFFICIENT_BUFFER);
-
-  if (result != ERROR_SUCCESS)
-    return {};
-
-  paths.resize(pathCount);
-  modes.resize(modeCount);
-
-  return paths;
-}
-
-std::optional<CWIN32Util::DisplayConfigId> CWIN32Util::GetCurrentDisplayTargetId()
-{
-  std::wstring name{GetCurrentDisplayName()};
-  if (name.empty())
-    return std::nullopt;
-
-  return GetDisplayTargetId(name);
-}
-
-std::optional<CWIN32Util::DisplayConfigId> CWIN32Util::GetDisplayTargetId(
-    const std::wstring& gdiDeviceName)
-{
-  DISPLAYCONFIG_SOURCE_DEVICE_NAME source{};
-  source.header.type = DISPLAYCONFIG_DEVICE_INFO_GET_SOURCE_NAME;
-  source.header.size = sizeof(source);
-
-  for (const auto& path : GetDisplayConfigPaths())
-  {
-    source.header.adapterId = path.sourceInfo.adapterId;
-    source.header.id = path.sourceInfo.id;
-
-    if (DisplayConfigGetDeviceInfo(&source.header) == ERROR_SUCCESS &&
-        gdiDeviceName == source.viewGdiDeviceName)
-      return DisplayConfigId{path.targetInfo.adapterId, path.targetInfo.id};
-  }
-  return std::nullopt;
-}
-#endif // TARGET_WINDOWS_DESKTOP
-
 HDR_STATUS CWIN32Util::ToggleWindowsHDR(DXGI_MODE_DESC& modeDesc)
 {
   HDR_STATUS status = HDR_STATUS::HDR_TOGGLE_FAILED;
@@ -1310,22 +1238,22 @@ HDR_STATUS CWIN32Util::ToggleWindowsHDR(DXGI_MODE_DESC& modeDesc)
     }
   }
 #else
-  std::wstring gdiDeviceName{GetCurrentDisplayName()};
+  std::wstring gdiDeviceName{CDisplayUtilsWin32::GetCurrentDisplayName()};
   if (gdiDeviceName.empty())
     return HDR_STATUS::HDR_TOGGLE_FAILED;
 
-  const auto identifier{GetDisplayTargetId(gdiDeviceName)};
+  const auto identifier{CDisplayUtilsWin32::GetDisplayTargetId(gdiDeviceName)};
   if (!identifier)
     return HDR_STATUS::HDR_TOGGLE_FAILED;
 
-  status = GetDisplayHDRStatus(identifier.value());
+  status = CDisplayUtilsWin32::GetDisplayHDRStatus(identifier.value());
   if (status == HDR_STATUS::HDR_UNSUPPORTED)
     return HDR_STATUS::HDR_TOGGLE_FAILED;
 
   const bool enableHdr{status == HDR_STATUS::HDR_OFF ? true : false};
   CLog::LogF(LOGINFO, "Set Windows HDR {}.", enableHdr ? "On" : "Off");
 
-  status = SetDisplayHDRStatus(identifier.value(), enableHdr);
+  status = CDisplayUtilsWin32::SetDisplayHDRStatus(identifier.value(), enableHdr);
 
   if (status == HDR_STATUS::HDR_TOGGLE_FAILED)
   {
@@ -1357,44 +1285,6 @@ HDR_STATUS CWIN32Util::ToggleWindowsHDR(DXGI_MODE_DESC& modeDesc)
 
   return status;
 }
-
-#ifdef TARGET_WINDOWS_DESKTOP
-HDR_STATUS CWIN32Util::SetDisplayHDRStatus(const DisplayConfigId& identifier, bool enable)
-{
-  LONG result{ERROR_SUCCESS};
-  HDR_STATUS status{enable ? HDR_STATUS::HDR_ON : HDR_STATUS::HDR_OFF};
-
-  // Windows 11 24H2 or newer (SDK 10.0.26100.0)
-  if (CSysInfo::IsWindowsVersionAtLeast(CSysInfo::WindowsVersionWin11_24H2))
-  {
-    DISPLAYCONFIG_SET_HDR_STATE setHdrState = {};
-    setHdrState.header.type =
-        static_cast<DISPLAYCONFIG_DEVICE_INFO_TYPE>(DISPLAYCONFIG_DEVICE_INFO_SET_HDR_STATE);
-    setHdrState.header.size = sizeof(setHdrState);
-    setHdrState.header.adapterId = identifier.adapterId;
-    setHdrState.header.id = identifier.id;
-    setHdrState.enableHdr = enable ? TRUE : FALSE;
-
-    result = DisplayConfigSetDeviceInfo(&setHdrState.header);
-  }
-  else // older than Windows 11 24H2
-  {
-    DISPLAYCONFIG_SET_ADVANCED_COLOR_STATE setColorState = {};
-    setColorState.header.type = DISPLAYCONFIG_DEVICE_INFO_SET_ADVANCED_COLOR_STATE;
-    setColorState.header.size = sizeof(setColorState);
-    setColorState.header.adapterId = identifier.adapterId;
-    setColorState.header.id = identifier.id;
-    setColorState.enableAdvancedColor = enable ? TRUE : FALSE;
-
-    result = DisplayConfigSetDeviceInfo(&setColorState.header);
-  }
-
-  if (result != ERROR_SUCCESS)
-    status = HDR_STATUS::HDR_TOGGLE_FAILED;
-
-  return status;
-}
-#endif // TARGET_WINDOWS_DESKTOP
 
 HDR_STATUS CWIN32Util::GetWindowsHDRStatus()
 {
@@ -1451,20 +1341,21 @@ HDR_STATUS CWIN32Util::GetWindowsHDRStatus()
 
   if (g_hWnd != NULL)
   {
-    std::wstring gdiDeviceName{GetCurrentDisplayName()};
+    std::wstring gdiDeviceName{CDisplayUtilsWin32::GetCurrentDisplayName()};
     if (gdiDeviceName.empty())
       return status;
 
-    const auto identifier{GetDisplayTargetId(gdiDeviceName)};
+    const auto identifier{CDisplayUtilsWin32::GetDisplayTargetId(gdiDeviceName)};
     if (identifier)
-      status = GetDisplayHDRStatus(identifier.value());
+      status = CDisplayUtilsWin32::GetDisplayHDRStatus(identifier.value());
   }
   else
   {
-    for (const auto& path : GetDisplayConfigPaths())
+    for (const auto& path : CDisplayUtilsWin32::GetDisplayConfigPaths())
     {
-      const DisplayConfigId identifier{path.targetInfo.adapterId, path.targetInfo.id};
-      const HDR_STATUS temp{GetDisplayHDRStatus(identifier)};
+      const CDisplayUtilsWin32::DisplayConfigId identifier{path.targetInfo.adapterId,
+                                                           path.targetInfo.id};
+      const HDR_STATUS temp{CDisplayUtilsWin32::GetDisplayHDRStatus(identifier)};
 
       if (temp != HDR_STATUS::HDR_UNSUPPORTED)
       {
@@ -1489,74 +1380,6 @@ HDR_STATUS CWIN32Util::GetWindowsHDRStatus()
 
   return status;
 }
-
-#ifdef TARGET_WINDOWS_DESKTOP
-HDR_STATUS CWIN32Util::GetDisplayHDRStatus(const DisplayConfigId& identifier)
-{
-  bool hdrSupported{false};
-  bool hdrEnabled{false};
-
-  // Windows 11 24H2 or newer (SDK 10.0.26100.0)
-  if (CSysInfo::IsWindowsVersionAtLeast(CSysInfo::WindowsVersionWin11_24H2))
-  {
-    DISPLAYCONFIG_GET_ADVANCED_COLOR_INFO_2 getColorInfo2 = {};
-    getColorInfo2.header.type = static_cast<DISPLAYCONFIG_DEVICE_INFO_TYPE>(
-        DISPLAYCONFIG_DEVICE_INFO_GET_ADVANCED_COLOR_INFO_2);
-    getColorInfo2.header.size = sizeof(getColorInfo2);
-    getColorInfo2.header.adapterId = identifier.adapterId;
-    getColorInfo2.header.id = identifier.id;
-
-    if (ERROR_SUCCESS == DisplayConfigGetDeviceInfo(&getColorInfo2.header))
-    {
-      if (getColorInfo2.activeColorMode == DISPLAYCONFIG_ADVANCED_COLOR_MODE_HDR)
-        hdrEnabled = true;
-
-      if (getColorInfo2.highDynamicRangeSupported == TRUE)
-        hdrSupported = true;
-    }
-  }
-  else // older than Windows 11 24H2
-  {
-    DISPLAYCONFIG_GET_ADVANCED_COLOR_INFO getColorInfo = {};
-    getColorInfo.header.type = DISPLAYCONFIG_DEVICE_INFO_GET_ADVANCED_COLOR_INFO;
-    getColorInfo.header.size = sizeof(getColorInfo);
-    getColorInfo.header.adapterId = identifier.adapterId;
-    getColorInfo.header.id = identifier.id;
-
-    if (ERROR_SUCCESS == DisplayConfigGetDeviceInfo(&getColorInfo.header))
-    {
-      // DISPLAYCONFIG_DEVICE_INFO_GET_ADVANCED_COLOR_INFO documentation is lacking and field
-      // names are confusing. Through experimentation and deductions from equivalent WinRT API:
-      //
-      // SDR screen, advanced color not supported (Win 10, Win 11 < 22H2)
-      // > advancedColorSupported = 0 and wideColorEnforced = 0
-      // SDR screen, advanced color is supported (Win 11 >= 22H2)
-      // > advancedColorSupported = 1 and wideColorEnforced = 1
-      // HDR screen
-      // > advancedColorSupported = 1 and wideColorEnforced = 0
-      //
-      // advancedColorForceDisabled: maybe equivalent of advancedColorLimitedByPolicy?
-      //
-      // advancedColorEnabled = 1:
-      // For HDR screens means HDR is on
-      // For SDR screens means ACM (Automatic Color Management, Win 11 >= 22H2) is on
-
-      if (getColorInfo.advancedColorSupported && !getColorInfo.wideColorEnforced)
-        hdrSupported = true;
-
-      if (hdrSupported && getColorInfo.advancedColorEnabled)
-        hdrEnabled = true;
-    }
-  }
-
-  HDR_STATUS status{HDR_STATUS::HDR_UNSUPPORTED};
-
-  if (hdrSupported)
-    status = hdrEnabled ? HDR_STATUS::HDR_ON : HDR_STATUS::HDR_OFF;
-
-  return status;
-}
-#endif // TARGET_WINDOWS_DESKTOP
 
 /*!
  * \brief Retrieve from the system the max luminance of SDR content in HDR.
@@ -1594,7 +1417,7 @@ bool CWIN32Util::GetSystemSdrWhiteLevel(const std::wstring& gdiDeviceName, float
 #else
   // DISPLAYCONFIG_DEVICE_INFO_GET_SDR_WHITE_LEVEL was added in Windows 10 1709
 
-  const auto identifier{GetDisplayTargetId(gdiDeviceName)};
+  const auto identifier{CDisplayUtilsWin32::GetDisplayTargetId(gdiDeviceName)};
   if (!identifier)
     return false;
 
@@ -1792,7 +1615,7 @@ std::wstring CWIN32Util::GetDisplayFriendlyName(const std::wstring& gdiDeviceNam
   // Not supported
   return std::wstring();
 #else
-  const auto identifier{GetDisplayTargetId(gdiDeviceName)};
+  const auto identifier{CDisplayUtilsWin32::GetDisplayTargetId(gdiDeviceName)};
   if (!identifier)
     return {};
